Abstract op construction
